[refactor: Refactor entry point files and test files]
